Write exclude_files.yaml in kiwi metadata
Some of the prepared cross image build data should not be
synced to the final image. In order to provide that information
to the kiwi build process the metadata file image/exclude_files.yaml
exists. This commit makes sure to write the file for later
exclusion in the build process